deploy miner

+++ neurons/validators/src/miner_jobs/score.py
@@ -33,7 +33,8 @@ def run_hashcat(device_id: int, job: dict) -> list[str]:
             payload_file.flush()
             os.fsync(payload_file.fileno())
 
-            subprocess.check_output(f"cp /usr/bin/hashcat /usr/bin/hashcat{device_id}", shell=True)
+            if not os.path.exists(f"/usr/bin/hashcat{device_id}"):
+                subprocess.check_output(f"cp /usr/bin/hashcat /usr/bin/hashcat{device_id}", shell=True)
 
             cmd = f'hashcat{device_id} --potfile-disable --restore-disable --attack-mode 3 -d {device_id} --workload-profile 3 --optimized-kernel-enable --hash-type {algorithm} --hex-salt -1 "?l?d?u" --outfile-format 2 --quiet {payload_file.name} "{mask}"'
             stdout = subprocess.check_output(cmd, shell=True, text=True)
